user_api/requirements.txt
appdirs==1.4.3
argon2-cffi==19.1.0
asn1crypto==0.24.0
attrs==18.2.0
black==18.9b0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
ecdsa==0.13
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jwcrypto==0.6.0
MarkupSafe==1.1.0
powerline-status==2.7
pyasn1==0.4.5
pycparser==2.19
python-dateutil==2.8.0
python-jose==3.0.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.2.18
SQLAlchemy-Utc==0.10.0
toml==0.10.0
urllib3==1.24.1
Werkzeug==0.14.1
